Initial project configuration of folder structure, build system, test system and CI/CD
Description
This pull request configured the initial folder structure of the project, its build systems alongside scripts used for compilation, a base for testing integrated into the build system, and, configured the CircleCI CI/CD pipeline for the project.

Build System
Folder Structure
The CI folder contains scripts and utilities that are regularly used by the CI/CD pipeline.
The scripts folder contains scripts related to various aspects of the build system and utilities for the developer.
The tests folder contains the testing CMake configuration, the source code for the testing frameworks used, and the actual project tests source code.
The src subfolder contains the actual test source code.
The ThirdParty subfolder contains the source code used to build the test framework used. In this case catch2
The src folder contains the C++ source code for the project as well as individual modules cmakelists.
The subfolder Engine contains the source code for the engine
The subfolder Common contains code that is shared between the engine and the editor
The subfolder Editor contains the code for the editor that will be used alongside the engine. For the foreseeable future, this will be used as a test entry point for the engine code.
The .circleci contains the CircleCI configuration files
Cmake Scripts
This project makes heavy used of Cmake targets and subdirectories to structure itself.

There is a global CMakelists at the root of the project, this script handles controlling if tests should be included in the build or not. Tests can be turned on and off by setting the value of the CMake option variable ENABLE_TESTS.

Inside of src is where most of the project subdivision happens, there is a single CMakelists at the root of the src directory which just includes the subdirectories for the engine and the editor.

The engine is a cmake target that is compiled into a library, it can be compiled either into a static library or a dll and this can be done via the cmake variable ENGINE_DYNAMIC_COMPILE. In the case the engine is compiled as DLL some extra compiler macro defines are set so that the source code can know to toggle on dll exports and imports depending on the target. Source files in the engine directory are automatically detected and added to build during compilation, so no changes are required to the cmake script. Files in the Common folder are also automatically added to the build for the engine.

The editor is compiled into an actual executable. Files in the editor directory and common directory are automatically added to the editor build.

Both the engine and editor are configured to output their respective build files into a directory called bin in the build directory.

The test cmake configuration includes the editor, engine, and common code automatically so that it can be easily accessed to build tests. There is automatic test discovery so that CTest can easily be used. Binary files are output into a folder called tests inside of the bin folder in the build directory.

Build Scripts
Build scripts are contained in the folder scripts. This folder is organized in compiler folders, one for clang and one for msvc. Each compiler folder has a build folder and a tests folder, used to build the project and the tests of the project respectively. Both the tests and build folder have the same layout, containing a Debug and Release folder used to build and run those configurations as well as a build and run file for the last known configuration for the system. All the scripts presented here output their result in build/compiler folder.

clean.bat - deletes the entirety of the build folder.
config-clang-tidy.bat - configures the project to run clang-tidy on it.
lint.bat - runs source code style checking on the project via clang-format.
open-vs.bat - open the project in visual studio via the generated msvc sollution file.
static-analysis.bat - run static analysis on the project via clang-tidy.
compiler/clean.bat - deletes the build folder associated with the compiler
compiler/cmake.bat - configures the build folder and cmake for the specified compiler
compiler/build/run.bat - runs the project in the last used configuration, recompiling if required.
compiler/build/build.bat - compiles the project in the last used configuration.
compiler/build/_run_no_compile.bat - runs the project in the last compiled configuration. Project recompilation will not happen with this script.
compiler/build/Configuraiton(Release or Debug)/build.bat - builds the project in the specified configuration.
compiler/build/Configuraiton(Release or Debug)/build.bat - runs the project in the specified configuration, recompiling if required.
compiler/tests/run.bat - runs the tests in the last used configuration, recompiling if required.
compiler/tests/build.bat - compiles the tests in the last used configuration.
compiler/tests/_run_no_compile.bat - runs the tests in the last compiled configuration. Project recompilation will not happen with this script.
compiler/tests/Configuraiton(Release or Debug)/build.bat - builds the tests in the specified configuration.
compiler/tests/Configuraiton(Release or Debug)/build.bat - runs the tests in the specified configuration, recompiling if required.
CI/CD Pipeline Structure
Linting stage -> static analysis stage -> Compilation Test Stage -> Unit Test Stage -> Release Candidate Published on Github

Linting Stage - Linting is run via clang-format.
Static Analysis Stage - Runs C++ static analysis via clang-tidy.
Compilation Stage
MSVC Debug Compilation - Compiles the project in the Debug configuration via MSVC
MSVC Release Compilation - Compiles the project in the Release configuration via MSVC
Clang Debug Compilation - Compiles the project in the Debug configuration via Clang
Clang Release Compilation - Compiles the project in the Release configuration via Clang
Unit Test Stage
MSVC Debug Unit Tets - Runs the test in Debug mode compiled by MSVC
MSVC Release Unit Tets - Runs the test in Release mode compiled by MSVC
Clang Debug Unit Tets - Runs the test in Debug mode compiled by Clang
Clang Release Unit Tets - Runs the test in Release mode compiled by Clang
Release Candidate Published to Github - Publishes a release candidate to the GitHub releases marked as pre-released. This stage of the CI is only run in commits to the dev branch.